requirements.txt
streamlit
openai
python-dotenv
pyrebase4
pandas
google-auth
google-auth-oauthlib
google-api-python-client